Problem Statement :
Mr. Rahul is driving his car from his home to the office with "X" liters of petrol. There are "N" number of petrol bunks between his home and office, with only a limited capacity of petrol, and each petrol bunk is located in a different place. For every one km, one liter of petrol will be consumed. So he fills up the petrol tank in each petrol bunk. Help him to calculate whether the remaining liters of petrol is enough to reach the upcoming destination. If the tank is empty before reaching the destination, print the message "You are not able to travel". Otherwise, If he reaches the office, print the remaining liters of petrol in the tank.