Array Balancing in C
Consider an array [a0, a1,a2,… an-1] of size n where n is an even number. An array is balanced if the sum of the left half of the array elements is equal to the sum of right half.
To balance an array, add a non-negative integer x to any array element. Find the smallest value of ‘x’ that makes the array balanced.

Input format :
The first line of the input consists of the value of n.
Next input is the array elements.

Output format :
The output prints the smallest value of 'x' that makes the array balanced.
If the array is already balanced, print it.
Refer sample input and output for formatting specifications.