Diamond Pattern C
Nirmala participated in an interschool competition. New innovative ideas with design are motivated for a start-up. To come up with a new idea, she thought to build a Diamond-Shaped Pyramid of Alphabets. Assist her in developing her innovative idea and putting it into action.
Input format :
The input consists of an integer.

Output format :
The output prints the required pattern.