SubString in C
Problem Statement
Raju is preparing questions for the internal test. While creating questions, he gave one to his son, Arun.
Raju has given the string S to his son Arun. Arun's task is to print the size of the longest possible substring that has exactly K unique characters.
For example, in the string aabacbebebe with K = 3, the substring will be cbebebe with length 7.The input string is "aabacbebebe". The unique characters in the string are: 'a', 'b', 'c', 'e'. The longest subsequence with at most 3 distinct characters is "aabacbe". It contains the characters 'a', 'b', and 'c', which is the maximum number of distinct characters allowed (k = 3). Therefore, the length of the longest subsequence is 7, which is printed as the output.

Input format :
The first line of input contains a string S.
The next line contains an integer K.

Output format :
The output prints the size of the longest possible substring that has exactly K unique characters.

Code constraints :
1<= K <=10

Sample cases:
Input 1 :
aabacbebebe
3
Output 1 :
7

Input 2 :
abbc
1
Output 2 :
2