Positive and Negative Occurrence in C
Problem statement
Get the integer array input from the user and, print all the pairs having a positive value and negative value of a number that exists in the array. We need to print pairs in order of their occurrences. A pair whose element appears first should be printed first.

Input format :
The input consists of the number of elements in an array in the first line.
Array elements in the second line are separated by space.

Output format :
Display the pairs, as a positive number followed by a negative number.

Code constraints :
N >0.

Sample test cases:
Input 1:
7
2 3 1 4 5 -4 -2
Output 1:
2 -2 4 -4