Star Pattern in C
Problem Statement
Sameera, a receptionist at the Regency Hotel, rearranges the flowers in the reception daily. Arjun, the manager of the same hotel, writes a program to display the flower arrangement on the reception TV screen. Today, Sameera arranged the flowers as shown in the sample output. Help Arjun write a program to display the pattern.

Input format :
The input consists of an integer.

Output format :
The output prints the required pattern.

Sample test cases :
Input 1 :
5
Output 1 :
* * * * *  * * * * *
* * * *      * * * *
* * *          * * *
* *              * *
*                  *
*                  *
* *              * *
* * *          * * *
* * * *      * * * *
* * * * *  * * * * *
Input 2 :
3
Output 2 :
* * *  * * *
* *      * *
*          *
*          *
* *      * *
* * *  * * *