Pattern2 in C
ndhu, a receptionist at Vijay Park Inn rearranges the flower in the reception daily. Vikram, manager of the same hotel writes a program to display the flower arrangement on the reception TV screen. Today, Indhu arranged the flowers as shown in the sample output. Help Vikram to write a program to display the pattern.

Input format :
The input consists of an integer.

Output format :
The output prints the required pattern.

Input:
5
Output:

*                 * 
* *             * * 
* * *         * * * 
* * * *     * * * * 
* * * * * * * * * * 
* * * *     * * * * 
* * *         * * * 
* *             * * 
*                 *