Unique in C
Use a single-subscripted array to solve the following problem. The program should read a series of 'n' numbers between 10 and 100 (inclusive). Each number that is read should only be printed if it is unique and has not been previously read.

Input format :
The first line of the input consists of the value of n.
The next line of input is the array elements.

Output format :
The output prints the unique elements in the array.

Sample test cases :
Input 1 :
12
23 45 69 87 45 63 22 88 23 69 44 52
Output 1 :
23 45 69 87 63 22 88 44 52