Multiplying and Arrange in Ascending order in C
Alice is working on an array manipulation program. She enters an array of integers and a multiplier. The program then multiplies each array element by the given multiplier, sorts the array in descending order, and outputs the result.
Write a program to solve the problem.

Input format :
The first line contains an integer n, representing the size of the array.
The second line contains n space-separated integers, the elements of the array.
The third line contains an integer mul, representing the multiplier.

Output format :
The output displays a single line containing n space-separated integers, the elements of the array, after multiplying each element by the given multiplier and sorting in descending order.

Code constraints :
1 ≤ N ≤ 10
1 ≤ elements ≤ 100
1 ≤ mul ≤ 20

Input 1 :
5
10 50 20 30 40
2

Output 1 :
100 80 60 40 20