Non-Repeating Element in C
Aakash is implementing a program that finds the first non-repeating element in an array of integers. 
The program should take an integer N as input, representing the size of the array, and then read N integers from the user as the array elements. Afterwards, it should identify and print the first non-repeating element in the array. If no such element exists, it should indicate that no non-repeating element was found.
Assist Aakash in the program.

Input format :
The first line of input consists of an integer N, representing the size of the array.
The second line consists of N space-separated integers representing the array elements.

Output format:
The output displays following format:
If a non-repeating element is found in the array, print the element.
Otherwise, print "No non-repeating element found."

Code constraints :
1 ≤ N ≤ 20
0 ≤ array elements ≤ 100

Sample test cases :
Input 1 :
6
1 4 2 5 1 2
Output 1 :
4
Input 2 :
8
1 2 3 4 1 2 3 4
Output 2 :
No non-repeating element found.